refactor: :truck: moved styles files into component/styles folder
