+++ src/components/bingo/BoardTask.tsx
@@ -1,7 +1,7 @@
 import { useEffect, useRef, useState, type ChangeEvent } from "react";
 import { fileToUri } from "../../utils/formatFiles";
 import type { Task } from "../../utils/shuffleTasks";
-import "./BoardTask.css";
+import "./styles/BoardTask.css";
 
 interface Props {
   task: Task;
