+++ src/components/bingo/Modal.tsx
@@ -1,4 +1,4 @@
-import "./Modal.css";
+import "./styles/Modal.css";
 
 interface Props {
   open: boolean;
