+++ src/components/bingo/UsernameInput.astro
@@ -1,0 +1,114 @@
+---
+
+---
+
+<form class="username-input-container">
+  <input
+    type="text"
+    name="username"
+    placeholder="Ingresa tu nombre..."
+    maxlength="20"
+    autocomplete="off"
+    required
+  />
+  <button type="submit" disabled>Guardar</button>
+</form>
+
+<style>
+  .username-input-container {
+    display: flex;
+    align-items: center;
+    justify-content: center;
+    width: 70%;
+    height: 100%;
+
+    margin: 2rem auto;
+    padding: 1rem;
+
+    border-radius: 2rem;
+    box-shadow: var(--shadow-md);
+
+    background: var(--color-bg);
+  }
+
+  .username-input-container input {
+    width: 100%;
+    height: 20px;
+
+    margin: 0;
+    padding: 0.5rem 1rem;
+
+    color: var(--color-text);
+    font-size: var(--text-size-sm);
+
+    border: none;
+    border-top-left-radius: 2rem;
+    border-bottom-left-radius: 2rem;
+    box-shadow:
+      2px 1px 1px hsla(0, 0%, 0%, 0.07),
+      2px 2px 2px hsla(0, 0%, 0%, 0.1);
+    outline: none;
+
+    transition: all 0.3s ease-in-out;
+    -webkit-transition: all 0.3s ease-in-out;
+    -moz-transition: all 0.3s ease-in-out;
+    background: var(--color-bg-light);
+  }
+
+  .username-input-container button {
+    margin: 0;
+    width: 40%;
+    height: 36px;
+    padding: 0.5rem 1rem;
+
+    font-size: var(--text-size-sm);
+    font-weight: 300;
+    color: var(--color-text-muted);
+
+    border: none;
+    border-top-right-radius: 2rem;
+    border-bottom-right-radius: 2rem;
+    box-shadow:
+      2px 1px 1px hsla(0, 0%, 0%, 0.07),
+      2px 2px 2px hsla(0, 0%, 0%, 0.1);
+
+    transition: all 0.3s ease-in;
+
+    background: var(--color-bg-light);
+    cursor: pointer;
+    animation: textPulse 2s infinite;
+  }
+
+  .username-input-container button:disabled {
+    color: transparent;
+  }
+</style>
+
+<script>
+  document.addEventListener("DOMContentLoaded", () => {
+    const $form = document.querySelector(
+      ".username-input-container",
+    ) as HTMLFormElement;
+    const $input = $form.querySelector("input") as HTMLInputElement;
+    const $button = $form.querySelector("button") as HTMLButtonElement;
+
+    $input.addEventListener("input", () => {
+      if ($input.value.trim() !== "") {
+        $input.classList.add("changed");
+        $button.disabled = false;
+      } else {
+        $input.classList.remove("changed");
+        $button.disabled = true;
+      }
+    });
+
+    $form.addEventListener("submit", (e) => {
+      e.preventDefault();
+      const $input = $form.querySelector("input") as HTMLInputElement;
+
+      if ($input.value.trim() === "") return;
+
+      // Envio a DB o LocalStorage
+    });
+  });
+</script>
