+++ src/components/bingo/styles/Board.css
