+++ src/components/bingo/styles/Modal.css
